// src/app/modules/mantenimiento/models/v-cliente.model.ts
export class Vcliente {
    idCliente: number = 0;
    nombre: string = "";
    apPaterno: string = "";
    apMaterno: string = "";
    celular: number = 0;
    email: string = "";
    documento: string = "";
    idTipoDocumento: number = 0;
    nroDocumento: number = 0;
    distrito: string = "";
    calle: string = "";
    numero: number = 0;
    referencia: string = "";
    contra: string ="";
}